+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.17] - 2024-02-19
+
 ### Fixed
 
 - Eliminate unuse date format code
@@ -98,7 +100,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Several UI fixes in next event, and add price format
 
-[unreleased]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.16...HEAD
+[unreleased]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.17...HEAD
+[0.0.17]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.16...0.0.17
 [0.0.16]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.15...0.0.16
 [0.0.15]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.14...0.0.15
 [0.0.14]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.13...0.0.14